add a reusable WalletNameForm
